Add LBR data to context for BPF_PROG_TYPE_PERF_EVENT progs
Last-branch-record is an intel CPU feature that can be configured to
record certain branches that are taken during code execution. This data
is particularly interesting for profile guided optimizations. perf has
had LBR support for a while but the data collection can be a bit coarse
grained.

We've (Facebook) have recently run a lot of tests experimenting with
filtered LBR data and feeding it to various PGO pipelines. We've seen
really good results (+2.5% throughput with lower cpu util and lower
latency) by feeding high request latency LBR branches to the compiler).
We use bpf to read a special request context ID (which is how we
associate branches with latency) from a fixed userspace address. Reading
from the fixed address is why bpf support is useful.

This patch adds support for LBR data to perf bpf progs.

Some notes:
* We use `__u64 entries[BPF_MAX_LBR_ENTRIES * 3]` instead of
  `struct perf_branch_entry[BPF_MAX_LBR_ENTRIES]` because checkpatch.pl
  warns about including a uapi header from another uapi header

* We define BPF_MAX_LBR_ENTRIES as 32 (instead of using the value from
  arch/x86/events/perf_events.h) because including arch specific headers
  seems wrong and could introduce circular header includes.