rust: Add kvmalloc() allocator type
The current global allocator is for `kmalloc()` family of functions.
Usually this is fine, but if we try to allocate a large structure on the
heap, `kmalloc()` might fail because there is not enough contiguous
memory available left in the system.

There is currently no way to use a different allocator if you know you
will need a large allocation but don't necessarily need it to be
contiguous memory. This is the motivation for this commit.

This commit adds the `KernelVirtualAllocator` type that implements the
`Allocator` trait for kvmalloc() family of functions.

Signed-off-by: Daniel Xu <dxu@dxuuu.xyz>