bpftool: Add SPDX identifier to btf-dump-file output
A concern about potential GPL violations came up at the new $DAYJOB when
I tried to vendor the vmlinux.h output. The central point was that the
generated vmlinux.h does not embed a license string -- making the
licensing of the file non-obvious.

This commit adds a LGPL-2.1 OR BSD-2-Clause SPDX license identifier to
the generated vmlinux.h output. This is line with what bpftool generates
in object file skeletons.