bpf, net, frags: Add bpf_ip_defrag() kfunc
When called, this kfunc does exactly one of the following:

1. Complete the original packet (using previously passed in frags),
   update the passed in skb, and return 0
2. Discover the frag is part of a yet-incomplete original packet, take
   ownership of the passed in skb, and return non-zero
3. Encounter an error and return non-zero

The verifier ensures that the prog checks the return code. Only in == 0
branches may the prog further access / reference the skb (context). In
all other branches the skb is consumed and thus, inaccessible.

So far this helper is only callable from TC clsact progs.