bpf: Teach verifier about MAY_CONSUME_CTX
This patch teaches the verifier to invalidate all ctx accesses after a
call to a MAY_CONSUME_CTX kfunc. Then once the return value on said
kfunc is checked, the verifier marks the branches as having their ctx
valid or not, as appropriate.

Note that we do not explicitly track which return values are associated
to which kfunc. The reason is two fold:

1. There is currently only one MAY_CONSUME_CTX kfunc, so there is not
   yet any ambiguity.

2. Calling a MAY_CONSUME_CTX kfunc requires accessing the ctx. Thus,
   only one MAY_CONSUME_CTX kfunc can be outstanding at any point in
   time b/c the ctx is invalidated between the kfunc call and the return
   value check.