bpf: verifier: Allow calling bpf_user_ringbuf_drain() on ringbuf
Implementing a correct and high performance MPSC ringbuffer is tricky,
so it's best to reuse battle tested code.  Both kernel-to-user and
user-to-kernel are already supported.  The final "quadrant" is
kernel-to-kernel. We ignore user-to-user as there's no need to involve
the kernel for that - plenty of libraries exist.

This commit relaxes the verifier to allow calling
bpf_user_ringbuf_drain() on BPF_MAP_TYPE_RINGBUF. This makes it possible
to both produce and consume samples on the ringbuf from BPF progs.