tools/build: Skip jobserver flgas in -s detection
Currently there is unnecessarily verbose output:

    $ make -j8 bzImage
    mkdir -p /home/dlxu/dev/linux/tools/objtool && make
      O=/home/dlxu/dev/linux subdir=tools/objtool --no-print-directory -C
      objtool
    mkdir -p /home/dlxu/dev/linux/tools/bpf/resolve_btfids && make
      O=/home/dlxu/dev/linux subdir=tools/bpf/resolve_btfids
      --no-print-directory -C bpf/resolve_btfids
      INSTALL libsubcmd_headers
      INSTALL libsubcmd_headers
      UPD     include/config/kernel.release

The reason this happens is that it seems that make is internally adding
the following flag to $(MAKEFLAGS):

    ---jobserver-auth=fifo:/tmp/GMfifo1880691

This breaks -s detection which searches for 's' in $(short-opts), as any
this entire long flag is treated as a short flag and the presence of any
's' triggers silent=1.

Fix by filtering out such flags so it's still correct to do a substring
search for 's'.