XXX: bpf: Move bpf_check_attach_target() to core
This routine is used in a few other places in BPF subsystem. In an
effort to make verifier.c a leaf node, move the definition as well as
some other single use helpers into core.

XXX: Unclear if this is the cleanest way to do this. Another option is
to have verifier.c export a function pointer. Callers can then
rcu_dereference_protected() the global hook, as the module has a refcnt
held already.

Signed-off-by: Daniel Xu <dxu@dxuuu.xyz>