XXX: bpf: Supporting building verifier.ko out-of-tree
This supports out-of-tree builds against the in-tree verifier.ko. This
is intended to be used to build newer upstream verifiers against older
kernel sources.

You can also do an "in-tree out-of-tree" build as proof of concept:

    $ cd kernel/bpf
    $ $ make -C ../.. M=$PWD

Note we use a standard technique introduced in a9a49c2ad9b9 ("kbuild:
use $(srctree) instead of KBUILD_SRC to check out-of-tree build") to
distinguish between in-tree and out-of-tree build.

XXX: There's a bunch of warnings like:
    WARN: resolve_btfids: unresolved symbol task_struct
when doing an in-tree out-of-tree build